+++ src/vdc.hpp
@@ -71,9 +71,8 @@ struct layer_t {
 	//             | | | |
 	//             | | | +- hidden (0) / visible (1)
 	//             | | +--- tilemode (0) / bitmapped (1)
-	//             | +----- The 0b00 pattern is opaque (0) or transparent (1)
-	//             +------- The 0b11 pattern  assumes color 3 (0) or color from memory (1) (color per tile)
-	//
+	//             | +----- The 0b00 pattern is opaque (0) / transparent (1)
+	//             +------- The 0b11 pattern assumes color 3 (0) or color from memory (1) (color per tile)
 	// -----------------------------------------------------------------
 	bool flags0_bit0_visible;
 	bool flags0_bit1_bitmapped;
@@ -84,11 +83,14 @@ struct layer_t {
 	// flags1
 	//
 	// bit 7 6 5 4 3 2 1 0
-	//     | | | |
+	//     | | | |     | |
+	//     | | | |     | +- flip h  (1)
+	//     | | | |     +--- flip v  (1)
 	//     | | +-+--------- horizontal stretch (0b00 = 1x, 0b01 = 2x, 0b10 = 4x, 0b11 = 8x)
 	//     +-+------------- vertical stretch (0b00 = 1x, 0b01 = 2x, 0b10 = 4x, 0b11 = 8x)
-	//
 	// -----------------------------------------------------------------
+	bool flags1_bit0_flip_h;
+	bool flags1_bit1_flip_v;
 	uint8_t flags1_bit45_hstretch;
 	uint8_t flags1_bit67_vstretch;
 
@@ -97,8 +99,8 @@ struct layer_t {
 	//
 	// bit 7 6 5 4 3 2 1 0
 	//         | |     | |
-	//         | |     +-+- hor size (0b00 = 4, 0b01 = 8, 0b10 = 16, 0b11 = 32)
-	//         +-+--------- ver size (0b00 = 4, 0b01 = 8, 0b10 = 16, 0b11 = 32)
+	//         | |     +-+- hor tile size (0b00 = 4px, 0b01 = 8px, 0b10 = 16px, 0b11 = 32px)
+	//         +-+--------- ver tile size (0b00 = 4px, 0b01 = 8px, 0b10 = 16px, 0b11 = 32px)
 	// -----------------------------------------------------------------
 	uint8_t flags2_bit01_hsize;
 	uint8_t flags2_bit45_vsize;
@@ -118,36 +120,27 @@ struct sprite_t {
 	// flags0
 	//
 	// bit 7 6 5 4 3 2 1 0
-	//     | | | |   |   |
-	//     | | | |   |   +- hidden (0) / visible (1)
-	//     | | | |   +----- 0b00 patterns code to opaque (0) or transparent (1)
-	//     | | | |
-	//     | | | +--------- x pos relative to screen (0) or associated layer (1)
-	//     | | +----------- y pos relative to screen (0) or associated layer (1)
+	//     | |       |   |
+	//     | |       |   +- hidden (0) / visible (1)
+	//     | |       +----- 0b00 patterns code to opaque (0) or transparent (1)
 	//     +-+------------- transparency of sprite (0b00 = no transp., 0b01/0b01/0b11 various degrees)
-	//
 	// -----------------------------------------------------------------
 	bool flags0_bit0_visible;
 	bool flags0_bit2_transparent;
-	bool flags0_bit4_xpos_rel_layer;
-	bool flags0_bit5_ypos_rel_layer;
 	uint8_t flags0_bit67_transparency;
 
 	// -----------------------------------------------------------------
 	// flags1
 	//
 	// bit 7 6 5 4 3 2 1 0
-	//     | | | |   | | |
-	//     | | | |   | | +- flip h  (1)
-	//     | | | |   | +--- flip v  (1)
-	//     | | | |   +----- flip xy (1)
+	//     | | | |     | |
+	//     | | | |     | +- flip h  (1)
+	//     | | | |     +--- flip v  (1)
 	//     | | +-+--------- horizontal stretch (0b00 = 1x, 0b01 = 2x, 0b10 = 4x, 0b11 = 8x)
 	//     +-+------------- vertical stretch (0b00 = 1x, 0b01 = 2x, 0b10 = 4x, 0b11 = 8x)
-	//
 	// -----------------------------------------------------------------
 	bool flags1_bit0_flip_h;
 	bool flags1_bit1_flip_v;
-	bool flags1_bit2_flip_xy;
 	uint8_t flags1_bit45_hstretch;
 	uint8_t flags1_bit67_vstretch;
 
@@ -156,8 +149,8 @@ struct sprite_t {
 	//
 	// bit 7 6 5 4 3 2 1 0
 	//         | |     | |
-	//         | |     +-+- hor size (0b00 = 4, 0b01 = 8, 0b10 = 16, 0b11 = 32)
-	//         +-+--------- ver size (0b00 = 4, 0b01 = 8, 0b10 = 16, 0b11 = 32)
+	//         | |     +-+- hor sprite size (0b00 = 4px, 0b01 = 8px, 0b10 = 16px, 0b11 = 32px)
+	//         +-+--------- ver sprite size (0b00 = 4px, 0b01 = 8px, 0b10 = 16px, 0b11 = 32px)
 	// -----------------------------------------------------------------
 	uint8_t flags2_bit01_hsize;
 	uint8_t flags2_bit45_vsize;
@@ -189,7 +182,7 @@ private:
 
 	void draw_scanline(uint16_t scanline);
 	void draw_scanline_layer(layer_t *l, uint16_t sl);
-	void draw_scanline_sprite(sprite_t *s, uint16_t sl, layer_t *t);
+	void draw_scanline_sprite(sprite_t *s, uint16_t sl);
 
 	exceptions_ic *exceptions;
 	sn74ls148_t *sn74ls148;